업비트 시세 통신 설정

+++ src/main/java/com/uj/stxtory/domain/dto/deal/DealInfo.java
@@ -52,7 +52,7 @@ public abstract class DealInfo {
                     prices = getPriceByPage(item, 1, getPage());
 
                     // 조회 기간(6개월) 중 신고가가 아니면 제외
-                    checkPrice = prices.parallelStream().max(Comparator.comparingLong(DealPrice::getHigh)).orElse(null);
+                    checkPrice = prices.parallelStream().max(Comparator.comparingDouble(DealPrice::getHigh)).orElse(null);
                     return checkPrice != null && prices.get(lastdayIndex).getHigh() == checkPrice.getHigh();
                 })
                 // 로그
