+++ src/main/java/com/uj/stxtory/domain/dto/deal/DealItem.java
@@ -9,7 +9,7 @@ public interface DealItem {
 
     String getCode();
     String getName();
-    long getExpectedSellingPrice();
-    long getMinimumSellingPrice();
+    double getExpectedSellingPrice();
+    double getMinimumSellingPrice();
     int getRenewalCnt();
 }
