+++ src/main/java/com/uj/stxtory/domain/dto/deal/DealPrice.java
@@ -3,10 +3,10 @@ package com.uj.stxtory.domain.dto.deal;
 import java.util.Date;
 
 public interface DealPrice {
-    long getVolume();
-    long getDiff();
-    long getClose();
-    long getHigh();
-    long getLow();
+    double getVolume();
+    double getDiff();
+    double getClose();
+    double getHigh();
+    double getLow();
     Date getDate();
 }
